Make test names consistent
Test functions are named test_filename__functionname(), where there is a
double underscore after the filename, before the name of the function to
be tested. If the function being tested is internal, and starts with an
underscore, that underscore is kept, giving a triple underscore in the
test function name.